Revert back to using bytes to represent characters

+++ org.spoofax.jsglr2/src/main/java/org/spoofax/jsglr2/actions/IAction.java
@@ -8,7 +8,7 @@ public interface IAction {
 
     ICharacters characters();
 
-    public default boolean appliesTo(int character) {
+    public default boolean appliesTo(byte character) {
         return characters().containsCharacter(character);
     }
